+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Support permission for loan-type maintenance. Fixes UIIT-38.
 * Rely on new-style permission name to guard material-types settings.Fixes UIIT-39.
+* Use `<ControlledVocab>` from stripes-smart-components instead of `<AuthorityList>` from stripes-components. See STSMACOM-6. Requires util-notes v0.3.0. Fixes UIIT-74.
 
 ## [1.11.0](https://github.com/folio-org/ui-items/tree/v1.11.0) (2017-09-01)
 [Full Changelog](https://github.com/folio-org/ui-items/compare/v1.10.1...v1.11.0)